% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimeo.R
\name{get_myVideoData}
\alias{get_myVideoData}
\title{Get information of all my videos}
\usage{
get_myVideoData(auth.code = vimeo_auth())
}
\arguments{
\item{auth.code}{A list, returned by vimeo_auth()}
}
\value{
A data frame.
}
\description{
Get information of all my videos
}
\examples{
none
}
